Add workaround to fix decoding of nested structs
Poison has a bug when using Poison.decode!/2 with automatic conversion
to structs that causes the Poison.Decoder protocol for the corresponding
struct to be called twice. What ends up happening with the DocuSign
models is that the second decode (using default values for the struct)
overwrites the ones coming for the response.

This workaround avoids calling the Poison.Decoder protocol a second
time.

For more information see: https://github.com/devinus/poison/issues/191