Avoid creating multipart request when only a single entity is sent
Previously, all the entities being sent as the request body were
enclosed in a Tesla.Multipart. DocuSign will apparently return a
content type error in the cases in which a single entity is expected
in the body.

These modifications were tested with the DocuSign.Api.EnvelopeViews
module, which was previously failing when using the
views_post_envelope_recipient_view/4 function.